reads are synchronous, re-org code
