+++ cabby.go
@@ -9,13 +9,19 @@ import (
 	"strconv"
 )
 
-const configPath = "config/cabby.json"
+const (
+	minBuffer          = 10
+	configPath         = "config/cabby.json"
+	stixContentType20  = "application/vnd.oasis.stix+json; version=2.0"
+	stixContentType    = "application/vnd.oasis.stix+json"
+	taxiiContentType20 = "application/vnd.oasis.taxii+json; version=2.0"
+	taxiiContentType   = "application/vnd.oasis.taxii+json"
+)
 
 var (
-	config   cabbyConfig
-	logInfo  = log.New(os.Stderr, "INFO: ", log.Ldate|log.Ltime|log.Lmicroseconds|log.Lshortfile|log.LUTC)
-	logWarn  = log.New(os.Stderr, "WARN: ", log.Ldate|log.Ltime|log.Lmicroseconds|log.Lshortfile|log.LUTC)
-	logError = log.New(os.Stderr, "ERROR: ", log.Ldate|log.Ltime|log.Lmicroseconds|log.Lshortfile|log.LUTC)
+	info = log.New(os.Stderr, "INFO: ", log.Ldate|log.Ltime|log.Lmicroseconds|log.Lshortfile|log.LUTC)
+	warn = log.New(os.Stderr, "WARN: ", log.Ldate|log.Ltime|log.Lmicroseconds|log.Lshortfile|log.LUTC)
+	fail = log.New(os.Stderr, "ERROR: ", log.Ldate|log.Ltime|log.Lmicroseconds|log.Lshortfile|log.LUTC)
 )
 
 func init() {
@@ -30,18 +36,16 @@ func newCabby() *http.Server {
 func registerAPIRoot(apiRoot string, h *http.ServeMux) {
 	u, err := url.Parse(apiRoot)
 	if u.Path == "" || err != nil {
-		logWarn.Panic(err)
+		warn.Panic(err)
 	}
 
-	logInfo.Println("Registering API handler for", u)
-	h.HandleFunc(u.Path, basicAuth(handleTaxiiAPIRoot))
-	h.HandleFunc(u.Path+"/collections", basicAuth(handleTaxiiCollection))
+	info.Println("Registering API handler for", u)
+	h.HandleFunc(u.Path, handleTaxiiAPIRoot)
+	h.HandleFunc(u.Path+"/collections/", handleTaxiiCollections)
 }
 
 func setupHandler() *http.ServeMux {
 	handler := http.NewServeMux()
-
-	handler.HandleFunc("/taxii", basicAuth(handleTaxiiDiscovery))
 
 	for _, apiRoot := range config.Discovery.APIRoots {
 		if config.validAPIRoot(apiRoot) {
@@ -49,18 +53,20 @@ func setupHandler() *http.ServeMux {
 		}
 	}
 
+	handler.HandleFunc("/taxii", handleTaxiiDiscovery)
 	handler.HandleFunc("/", handleUndefinedRequest)
 
 	return handler
 }
 
+// server is set up with basicAuth and HSTS applied to each handler
 func setupServer(h http.Handler) *http.Server {
 	port := strconv.Itoa(config.Port)
-	logInfo.Println("Server will listen on port " + port)
+	info.Println("Server will listen on port " + port)
 
 	return &http.Server{
 		Addr:         ":" + port,
-		Handler:      hsts(h),
+		Handler:      basicAuth(h),
 		TLSConfig:    setupTLS(),
 		TLSNextProto: make(map[string]func(*http.Server, *tls.Conn, http.Handler)),
 	}
@@ -82,5 +88,5 @@ func setupTLS() *tls.Config {
 
 func main() {
 	server := newCabby()
-	logError.Fatal(server.ListenAndServeTLS(config.SSLCert, config.SSLKey))
+	fail.Fatal(server.ListenAndServeTLS(config.SSLCert, config.SSLKey))
 }
