+++ taxii_error.go
@@ -1,0 +1,11 @@
+package main
+
+type taxiiError struct {
+	Title           string            `json:"title"`
+	Description     string            `json:"description,omitempty"`
+	ErrorID         string            `json:"error_id,omitempty"`
+	ErrorCode       string            `json:"error_code,omitempty"`
+	HTTPStatus      int               `json:"http_status,string,omitempty"`
+	ExternalDetails string            `json:"external_details,omitempty"`
+	Details         map[string]string `json:"details,omitempty"`
+}
