+++ taxii_status.go
@@ -1,0 +1,14 @@
+package main
+
+type taxiiStatus struct {
+	ID               taxiiID  `json:"id"`
+	Status           string   `json:"status"`
+	RequestTimestamp string   `json:"request_timestamp"`
+	TotalCount       int64    `json:"total_count"`
+	SuccessCount     int64    `json:"success_count"`
+	Successes        []string `json:"successes"`
+	FailureCount     int64    `json:"failure_count"`
+	Failures         []string `json:"failures"`
+	PendingCount     int64    `json:"pending_count"`
+	Pendings         []string `json:"pendings"`
+}
